MaaS testing
Currently there is no regular testing of the MaaS playbooks. This
commit attempts to address that.

generate-definitions.py is used to create files that define the MaaS
configuration for a set of specified entities.

The files in maas/testing/tests/integration are per-MaaS-playbook
reference definitions files and were created using
generate-definitions.py. These need updating when the MaaS configuration
is altered.

compare-definitions.py is used to compare the definitions defined in the
this repo with those generated for the deployment being tested.

generate-docs.py prints tables of the checks, alarms and metrics
defined in one or more definitions files. The format of the output is
reStructuredText. To print the tables for the definitions set:

  ./generate-docs.py tests/integration/*

The test-maas.yml playbook can be used to test a deployments against
the reference definitions files. It prepares the hosts for testing and
then uses generate-definitions.py and compare-definitions.py to test
the deployed MaaS configuration against the one found in
maas/testing/tests/integration. As part of the setup a script is copied
to the hosts to allow HP hardware testing to be configured. Obviously
this does not test that the plugin works with the HP binaries, but that
that checks/alarms are configured as expected.

Closes-issue: https://github.com/rcbops/rpc-openstack/issues/216
(cherry picked from commit 738440d3ea06968e9bd6c5f6cbe1a1e5422a8d95)