cinder_service_check.py not using --host option
The script cinder_service_check.py defines an optional host parameter
that is designed to filter the results, returned by the cinder API, so
that only services on the specified host are returned. The MaaS setup
role, that configures the checks that use this script, uses '--host'.
The script does not use the host argument if supplied.

This patch fixes the script so that the results returned are filtered
by host, where host includes any backends/pools on that host.

Closes-bug: https://github.com/rcbops/rpc-openstack/issues/200
(cherry picked from commit 6c4d898591bdf2cfdeed5c8d564dda622252daf6)