Fix MaaS testing
Currently the MaaS testing does not form part of a gate, this has caused
the testing code to diverge from the deployed monitoring. This commit
addresses a number of identified issues.

The generate-definitions.py script has been updated to allow specific
attributes to be templated. This is required to allow a password in the
RabbitMQ check to be replaced following an update to how that check is
defined.

Rackspace Cloud servers now automatically get a ping check created. The
test-maas.yml playbook will delete any existing checks from the
entities.

MaaS remote checks have a new attribute called collectors. The
collectors attribute will be removed because it does not need verifying.

The commit 1f7af9d4cfd909bd2ee5daa367d5d6fb8f468746 has broken the
script fake_hp_monitoring.py by removing a ')' from one of the strings.
This is re-added to resolve the issue. This script has also been updated
to allow the new metrics added to hp_monitoring.py to be tested.

The check neutron_metadata_agent_proxy_check has been added, this commit
updates the testing to support it.

Closes-Issue: https://github.com/rcbops/rpc-openstack/issues/351