WIP Add maas check validation
Add task outline for validating checks after their templates have
been processed and the raxmon agent has been restarted.

Challenge to be resolved now is that we would like to do this once
per entity (host) but we don't have a complete list of checks we are
looking for.

The best bet may be to do some processing on an ls -1 of the dest
directory in a shell to extract the filenames portion before '--' and
use that as the set to search for in the checks list generated by the
rax-maas-tool.py.