Upgrade maas to use template configuration
Adds a playbook to be run as a one-time operation to convert maas to
use template based configuration. The playbook records a fingerprint
of when it runs under /etc/openstack_deploy/migrations and will not
run again if this file is present.

The play will generate .tab files on the deploy host under /tmp for
reporting the state of checks before and after upgrade.

Also fixed verify-maas play to use dash instead of underscore in
filename and made appropriate change to README to reflect that.