Allow linting.sh to exit on first failure
Using the trap to keep the linting script running beyond the first
failure was intended to provide complete coverage on each run of the
linting job, but in hindsight it buried information. The output of the
Ansible commands is long, resulting in a lot of output to parse.

Also, flake8 failures were being eaten by the `|| failed=1` logic, but not
being repeated at the end, so developers ended up having to re-run the
checks locally anyway to find out what exactly failed.