Use existing vars file in deploy.sh
The rpc deploy script currently has a checks whether some variable files
exist in /etc/openstack_deploy. If they don't the repo version is copied
in. The file is then merged with the repo version via update-yaml, the
repo values taking precedence.

IMHO this is backwards, user provided values should override values
provided in the repo. There is a larger problem though, which is that
this code for checking/updating vars files is only run if
/etc/openstack_deploy doesn't exist, so there will never be any user
specified files to read/update. This is probably why the reverse
precedence hasn't been detected.

What this patch does:
  * Check for the existence of
    /etc/openstack_deploy/openstack_user_config.yaml rather than
    /etc/openstack_deploy when deciding whether to bootstrap an aio.
    This allows a user to precreate that directory and supply their own
    vars files before running deploy.sh

  * Flip the precedence so that user provided values override repo
    provided values

Related: #767