pin setuptools to 19.3
this commit poins setuptools to 19.3 in the galera containers.
setuptools>=19.4 does not allow holland installs to work the way we are
currently doing them (local pip installs).

This will unblock the kilo gate and allow us to work on the better
solution of having the holland pip wheels get built by the repo builder
and install that way, rather than a local pip install.