Add constraints file for all Pip packages
Add a constraints file generated by the repo server of an AIO build.

The repo server builds a single set of Python wheels based on the
required packages. These requirements vary in how specific a version
they want of any individual package. The result is that two deployments
build using the same configuration from the same SHA may end up with
different versions of wheels. Generally this does not cause an issue
however sometimes a new version of a Pip package is released that is
incompatible with the others required or has other dependencies beyond
what Pip can satisfy.

These incompatible package versions can cause testing/deployment
failures of releases and release candidates. This slows down the release
of new versions of this project and impacts customers. The constraints
file added by this commit is designed to more tightly control the wheels
that are build and so reduce the number of deployments that are impacted
by new versions of Pip packages.

The addition of this file creates additional overhead for the
development team, it should be regenerated for each new release. There
is also the possibility that it will impact a deployer where they
additional Python packages to be installed that are not compatible with
the constraints imposed by global-requirement-pins.txt. In this
situation it may be necessary for the deployer to modify or remove the
file.

The source file is found on the repo server of a sucessful build,
/var/www/repo/os-releases/12.0.8/requirements_absolute_requirements.txt.
Note this OSA generation of this file has a bug which means that the
line 'functools32==3.2.3_2' was replace with 'functools32==3.2.3.post2'.

Related-issue: https://github.com/rcbops/rpc-openstack/issues/910