Make Neutron use internal Nova endpoint
Update Neutron SHA and add configuration override to enable Neutron to
use Nova's internal endpoint.

This removes the need to change the OpenStack-Ansible SHA to address the
issue but will require an upgrade task to remove these changes in a
future release.

neutron_git_repo is required so that the lookup filter py_pkgs knows to
look in user_variables.yml for the new SHA.

Closes-issue: https://github.com/rcbops/rpc-openstack/issues/897