[WIP] Run ceph roles serially
This commit uses ansible's `serial` to deploy MONs and OSDs serially,
which should prevent masses of services restarting at the same time.

In order for MONs to bootstrap one by one, we need to set `mon initial
members` in ceph.conf.

Additionally, we separate out the tasks in ceph-mon.yml so these can
be targetted at a different host group (with the end goal of these
tasks only being run once).

Lastly, we add `max_fail_percentage: 0` to the deploy of ceph-osd so
that a failure will prevent the playbook from continuing.

NOTE: Bootstrapping OSDs serially may be slow, so we can introduce a
      variable override that can be passed in on the command line to
      disable `serial: 1`.