Add bashate linting
This commit adds the linting-bashate.sh script which is used to do
syntax checks on our scripts. Some important pieces of code are
scripts and previously our linting only checked YAML and Python
files. Upstream Openstack-Ansible also already includes bashate.

The linting-bashate.sh script itself searches recursively through
our files and runs bashate on files that have a (ba)sh shebang at
the top of the file. The script ignores the same bashate errors as
upstream OSA. The tox.ini was modified in order to add the
bashate script to the list of tests run. The .travis.yml file was
modified in order to add the bashate tox enviroment.

Addresses #454