Add the RPC plays for upgrading
Take the playbooks that were called as part of the old rpc-upgrade.sh
script and add them to the run-upgrade.yml playbook as includes.

Additionally, there is a final task that was run as part of the script
that should now run against "hosts" after the logging is setup.

Connects https://github.com/rcbops/u-suk-dev/issues/298