Remove Holland from outside venv
In mitaka, holland gets installed inside a venv, which means that we end
up with lingering holland pip packages outside the venv. To remove any
confusion, here we remove those packages as part of the post-upgrade.

Note, the pattern of 'check if holland is installed, then remove the pip
package if it is' is used only because in ansible 1.9.4, the pip module
is broken and it won't just "do the right thing" if you declare the
package as absent (it will actually fail with an rc of 1 if the package
has already been uninstalled)

Connected https://waffle.io/rcbops/u-suk-dev/cards/579f4f206d2c9e0a01448c6e