Do not install binary pycparser packages
pycparser's 2.14 wheel package on PyPI is broken and has been causing
breakage with cryptography and OpenStack in general. For rpc-openstack
in particular, ansible relies on cryptography and that dependency adds a
dependency on pycparser which has been breaking our gating.

Further, to avoid the binary packages we have to ensure pip is upgraded.

Connects rcbops/u-suk-dev#474