Workaround MaaS failure to properly update checks
When upgrading from Liberty to Mitaka the MaaS templates are updated.
The agent appears to be failing to correctly register the modifications
to the checks. Restarting the agent a second time appears to resolve the
issue.

This commit modifies the MaaS role so that the agent is restarted twice.

Connected https://github.com/rcbops/u-suk-dev/issues/481