Provide nova.yml env.d file for override
Per https://github.com/rcbops/u-suk-dev/issues/749
The default nova.yml file includes the compute hosts in the
neutron_l3_agent and neutron_metadata_agent group. This caused
our neutron checks to fail, since they use these groups to check
to see if the hosts in said groups have the l3 and metadata agent
running. Since those agents are not running on the compute hosts,
it will fail.
The newly supplied nova.yml is exactly the same as default,
except that it removes neutron_l3_agent and neutron_metadata_agent
as a group belonging to the compute hosts.