Set RPC-O to use the apt artifacts by default
This commit sets RPC-O to use the apt artifacts on rpc-repo
by default. The version is tied to rpc_release to ensure
that as each tag is released, newer artifacts are used.

Included in this commit is a clean-up of the artifacts
building variables to only include what it needs to complete
the container build process.

The original var names used in the artifact building process
have been adjusted in the implementation for RPC-O defaults
in order to reduce the possibility of namespace clashing.