Prevent race condition with LB
We download from LB, patch the file, then upload it.
Then for the next image, we do the same.
If the upload didn't got the chance to be rsynced across all nodes
with lsyncd, there is a chance that the file we will get will not
have all the latest changes.