Use recent OSA key/repo management
The pip_install role added the key/repo management to be
consistent everywhere (when doing RDO/UCA). A patch was also
included in OSA to be able to define an alternative repo to UCA.

We can now in RPC leverage this feature, and therefore we need
to wire the variables to point to our mirror that holds the UCA
packages.

This should do the trick.