Revert to using http for rpc-repo
There appears to be an issue with using https to rpc-repo
which results in the get_url tasks failing.

This patch reverts the setting to use http to bypass that
issue, unblocking other work that is ongoing.

The license boilerplate is also added and some keys are
cleaned up in the vars file used for the container artifact
build.

Connects https://github.com/rcbops/u-suk-dev/issues/1294