Make aptly use the first key only
When multiple key ids are listed, just import the top one.

Aptly roles uses the aptly key id as basis to know if the key
needs to be imported or not.

We are using a var for that, and that var is dynamically fetching
the id from the key that is in the CI system.

That should be enough for now.