Restrict Ubuntu updates for build/deploy
For the deployment to be repeatable it is crucial that
the apt sources do not include -security, -updates,
-backports, etc. All this must be provided through the
source added for the apt artifacts repo. For the build
and deployment to match in packages used, the build must
also be restricted.

Hosts and containers must only be enabled for the
'trusty' / 'xenial' repository which should, by default,
use the rackspace mirror. This mirror should be
unchanging after the initial distribution release.

For deployments in CDC's or in other environments with
a local mirror, we detect the existing host repository
and use it for the deployment.

Connects https://github.com/rcbops/u-suk-dev/issues/1310