Prevent the timeout of build
We need to report that something is happening during the build,
therefore we are using an async "fire and forget" job, and
regularily test if it's finished. The test should output the lines
we need to not timeout. When the task is finished, then we
can continue the play forward.