Refactor aptly to use env vars
This would make possible in CI system to trigger parts of the
process, depending on the need. The default is to run everything.

The idea is we could have a functional test that downloads,
updates repositories, create snapshots, publish them locally
on a throwaway node to exercice the building process, and
not push at the end of the process to rpc-repo.