Downgrade LXC cache packages when necessary
When we build python/container artifacts we pull the default
LXC cache variant from LXC upstream which may have packages
installed which are more recent than those we have available
in our artifacted apt repository. Here we figure out which
packages are installed, but not available in a configured
source, then we downgrade them to the latest available in
the configured sources.