Fixing the venv fetch path
The variable to download the venv is overriden in a group_var
per role. Because we don't add temp containers to group and
setup to get their vars, we override the vars ourselves.

This commit ensure we override the url for the venv fetch to
http://rpc-repo.rackspace.com/venvs/r14.0.0rc1/ubuntu-14.04-x86_64/
instead of
http://rpc-repo.rackspace.com/venvs/r14.0.0rc1/ubuntu/ like it
was before the SHA bump that changed the repo build behavior.