Purge ceph from RPC-O
This change removes the ceph installation framework we have in rpc-o.
This is being done because future ceph work will be owned and operated
by a different team. Furthermore, if ceph is needed to be deployed by
the RPC-O team we have ceph integrations being developed and maintained
upstream within OpenStack-Ansible.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>