Add playbooks and roles to deploy log aggregation container
This change is for issue/942 with the overall goal of filtering
and pushing customer logs from the local elk stack to a shared
global elk cluster.  This should give support easier access to
search customer logs for a quicker resolution with customer issues.
This is done in two roles with a playbook for each.  One takes
the elixir based aggie code, builds it in a local lxc continer,
packages it up and copies it to the local repo servers. The second
playbook will set up an lxc container created through openstack-ansible's
setup-hosts playbook and deploy the aggie code to it.  Where
the roles are in their own repo's, future changes should only
require a version update in the ansible-role-requirements.yml file
for anxible-galaxy.

Connects https://github.com/rcbops/u-suk-dev/issues/942