Updates host filter
Adding neutron_all group for post upgrade tasks includes the AIO
host. This results in the hostname of the AIO, which will eval to
true on `inventory_hostname in groups['utility'][0]`. This is a
problem as it will result in commands being run on the host rather
than in a utility container for post-upgrade checks. These fail
due to not having an openrc or openstack-client installed. The
updated host filter causes the openstack_services check to only
run on the expected utility node.