UG-621 Make staging work for multi-node builds
Currently the staging process only works when the deploy
host and the first infrastructure host are the same.

This patch implements a change to the process to still
allow a staging host to be specific manually, but it will
automatically use the inventory to identify the host for
the first repo server which is where the staging should
be done to.