Remove hard coded cinder backup enablement
This is being removed because it has been determined by the gating team
to be broken. Sadly the gating team has hard coded an additional
override for this override, which disbales this feature, so this code
path is never exercised in or outside of the gate. This removal will
allow us use well defined, and known overrides within the project
giving developers a clear sense of what is being developed vs what
is being tested and will unblock work in master.

Related-Issue: rcbops/rpc-gating#224
Un-Blocks: #2296

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>