RE-144 Fix OSA path in ansible wrapper
Currently, after a leapfrog upgrade is done, executing the ansible cli
under /opt/rpc-openstack/openstack-ansible/playbooks does not source
the ansible RC file, which results in the inventory not being found.
This is because the leapfrog upgrade modifies the ansible wrapper
pre-upgrade, but does not reset it back after the upgrade is done.

This commit simply resets the OSA path in the ansible wrapper post
upgrade.