This removes redundant scripts and update linters
This change beings using the tests repo from osa to better leverage well
curated and functional linters so that we're not having to mantain code
that is more/less a duplication to what we already have upstream. This
update follows the patterns we've used in RPC-MaaS ensuring good code is
getting written without the need to carry nested scripts.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>