RO-2429 Remove neutron_api_workers override
This patch removes the neutron_api_workers override that set the
number of API workers to 64. OSA already has code that calculates
an optimal number of neutron api workers by taking the number of
CPU cores and dividing by 2. If that value is incorrect, we should
change the logic in OSA.