RO-2431 Remove rpc_response_timeout overrides
In the Havana/Grizzly releases, RPC thread response timeout overrides
were required because of an issue with iptables when lots of
simultaneous builds were occurring. This was fixed in the Havana
release and these overrides are no longer needed.

The OSA default of 60 seconds is strongly recommended by the
oslo.messaging project. This patch removes the override so that the
safe default can be used.

Upstream Bug: https://bugs.launchpad.net/nova/+bug/1199433
Upstream Fix: https://review.openstack.org/#/c/40688/