RO-2428 Bump RabbitMQ role SHA
This patch provides some temporary relief from RO-2428 until there is
an upstream release of OpenStack-Ansible's Newton branch that
includes the fix.

This also brings in the fix for RO-2438.

This patch includes a backport of two master patches that allow
artifacting to add the upstream Erlang repository (part of RE-174):

  64edf05fc5f253ad21f654da4c56318e450016af
  a9cccdf3e0ae20880370f34e0a6971aa5e259c56