RE-674 Downgrade pkgs to apt artifact versions
When using images which have unattended upgrades
enabled, or have been built using packages which
are more recent than the available apt artifacts
the builds fail in strange ways.

This patch implements a function which will try
to downgrade any packages installed on the
deployment host and not available in the apt
artifacts to the versions which are.