RE-1082 Track MaaS tags
This PR adds MaaS version update to the update_dependencies hook script.
This means that the MaaS version RPCO uses will track MaaS
releases. If a new MaaS version is available, a PR will be created,
this PR will cause the new version of MaaS to be tested against RPC.