RO-2864 Remove a-r-r specs included in OSA
These role SHA's are now included in OSA, so we
should not override them any more.

Due to RO-3006 the galera_client and galera_server
role pins are left in place but updated to the last
SHA before the braking change. Once RO-3006 is
properly addressed these can be removed.