Set a package pin for the RPC-OpenStack repo
While the RPC-OpenStack repo "should" be the only repository used within
an artifacted deployment there are times when "other" repositories may
be added to a deployment post build which could have adverse affects on
our artifacted builds when it comes to day to operations. This change
adds a package pin instructing apt to prefer packages from our repos.
This change implements a pin using the highest priorty possible without
also instructing the APT to downgrade packages by default. The pin will
only be implemented when apt package artifacts are enabled and found to
be available.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>