Disable container image artifacts
This disables container/py artifacts for now. The RPC-OpenStack container
and py artifacts are failing import while the upstream container images
and builds of similart sizes, packages, and distros are not showing the
same issues. We need to spend some time debugging how the images are
built and how we can better construct / consume machine images.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>