RO-3068 override glance_default_store param
This commit overrides the value of the `glance_default_store` ansible
variable by injecting a value of `rbd` into the
`/etc/openstack_deploy/user_variables.yml` file as part of the deploy.

The `glance_default_store` ansible parameter is managed in multiple
locations which causes expected value of `rbd` to be replaced by a
value of `swift` by the file loading precedence. This change ensures
that the deploy script populates the desired value.