RO-3239 Ensure that RPC_RELEASE is correct
With the implementation of c4132cd the use of the
release.yml file from group_vars is no longer an
appropriate way to derive the RPC_RELEASE version.

This patch ensures that the RPC_RELEASE version is
correctly derived using a small python script in
order to read the release mapping.