RO-3239 Derive RPC_RELEASE correctly for imaging
The imaging process currently uses its own method to
derive the value for rpc_release. This method no longer
works since c4132cd merged.

Instead of uding its own method, given that it sources
the common functions, it already has an environment
variable available with the right version, so we use
that instead.