RO-3269 Remove ineffective pkg_locations override
The group_var to override the pkg_locations has no effect, resulting
in any values set there being ignored. It has no effect because OSA
Pike has this implemented as a playbook var, and that has a higher
precedence than a group_var.

The variation in the var from the upstream default is the addition of
the "/opt/rpc-openstack/rpcd" directory, which does not exist. Given
that we no longer have roles in-tree like we did with newton, this
override is necessary any more.

This patch therefore removes it.