RO-3268 Use apt artifacts for bootstrap-ansible
When install.sh runs it uses the bootstrap-ansible script from
OSA to install packages which are needed on the deployment host.
If the apt artifacts are not used for this, the host may end up
with packages which are newer than those available in the apt
artifacts and this will result in a failure later in the build.

This patch ensures that the availability of apt packages is
checked, and if they are available it uses them. There is also
an environment variable to override this behaviour as may be
necessary (for example, this is necessary when preparing the apt
artifacts).

The 'Remove extra sources' task in configure-apt-sources.yml is
removed as the task 'Replace the apt sources file with our content'
lays down a replaced file and the remove task therefore is always
a no-op.