RE-1309 Ensure update_dependencies script works
Currently if the update_dependencies script is run from a
directory which is not /opt/rpc-openstack, the get-rpc_release.py
script returns an empty result. This causes the version comparison
between the RC branch the the current development head to always
fail, so the version is never incremented.

This patch ensures that the RPC_RELEASE value is always properly
returned, and also ensures that the update_dependencies script is
easily tested on a test VM. Various updates are made to the script
to make it easier to re-run and to make it more robust.

The extended explanation for skipping the version bump is adjusted
due to bashate being incapable of understanding that the text using
the word "for" at the start of a line does not mean it's a bash for
loop.

Adjustments are also made to other gating scripts which now fail due
to the new verification being done.